feat: implement adding user to game room
